chore: use dedicated constants file for props/methods
